Change set state actions to support expressions
Rather than SET machine TO SYMBOL or SET machine TO PROPERTY symbol, this command
is now interpreted as SET machine TO expression. This is to make use of
the new PROPERTY SYMBOL and PROPERTY SYMBOL OF SYMBOL syntax that is
built into expressions.

See tests/prop.lpc for some example usage of properties, including a
couple of examples of the new implementation.